First commit of working server and handler
A lot of pieces had to be worked out to get this working, so this is a
big commit. I'd like to not make a habit of this, but I need something
working for my own websites.

This commit includes a server that sends CORS headers, support for
multiple handlers for a single path, plus logging to files and
stderr/stdout. Dependencies have been updated, and things should work
well overall. Some errors may be unclear, though, and this should be
considered a very beta program/library.