Live reload on config file changes
This commit adds a file watcher from the fsnotify project to watch for
changes to the configuration file and reload the configuration struct
and http server. It is done in such a way that connections should not be
dropped, though there's still the possibility of such an occurrence.

For library usage, the configuration file needs to be specified to the
file watching function, so the application can receive the path of the
changed file and handle the changed configuration itself (in case it's
not a TOML file).