Make more clear that the server started, not hung
Even though the server gets started after the message is printed, it
will have correctly started unless an error is returned, in which case
we'll have an error visible. The message "start*ing* server" made it
look like the server was hanging, not correctly working. This small edit
makes it a little more clear.