Prevent reassigning string value
Occasionally, the path for a handler would be replaced with one of the
other paths while logging the value inside the request handler function.
Now an unexported function generates the handler function, enclosing the
value of the path so references don't change. Also, VSCode changed tabs
and I'm not going to bother changing them back.