Resolve #1 and #4 - errors and multivalue forms
Overhauled the error handling portion of the code to pass a custom error
type, HTTPError, around instead of the built-in `error` type. HTTPError
bundles an HTTP Status code with the error, so the server knows which
status to respond to the client with.

If multiple errors occur, there is logic to choose the most
important/direct status and return that to the client.

Included an `Errorf` function in templates to enable form creators to
return custom error messages to users, for instance to prompt proper
formatting of a field.

As a small part of improving template functions, I also included a
FormValues (plural) function that returns a slice of string values
representing the multiple values returned from a form item, like
multiple checked checkboxes.