Resolve #5 - Implement honeypot spam detection
Form creators can specify a "honeypot" key in their handler's
configuration with a value matching the `name` attribute of the hidden
input field. Each handler checks their own honeypot field, so it is
possible to have multiple handlers for the same form but with different
honeypot fields.

If a honeypot is triggered, it is treated like a successful handling of
the form to avoid tipping off the spam bots.