Resolve #3 - conf.d support
static-forms now supports an `include_dir` configuration key that
specifies a location in which to look for more configuration files.
Configuration options are layered on top of each other, with the last
one specified taking precedence.

If there are two files in "conf.d" (the folder does not need to be
called this) and both have their own `include_dir` key, the included
files of the first file are parsed before the second file and its
included files.

The interparser dependency had to be updated to ensure errors were not
being overlook while doing this.