Move file watching into ParseConfigTOMLFile
Putting file watching here ensures that the file watcher is enabled for
every partial configuration file, such as those in a "conf.d" folder.
The channel still returns the file that was changed, but now `main()`
always loads the original file, found at `Config.RootConfig`.

This also seems to fix an issue where the file watcher stops if an error
occurs while reloading the config (and the old config is still used).