Updated SMTP Tests
Removed non-implemented test for sending emails, since that is largely
testing a dependency instead, and there is no way to really test if the
email was actually sent or if errors are due to code or the SMTP server.

Made SMTP configuration public (maybe) for use in other tests. Not sure
if I can access it from other packages, but it is nice to have. `gofmt`
was run on this file.